Fixing code smell. Making util file
