+++ frontend/src/containers/App/App.js
@@ -22,7 +22,6 @@ const App = () => {
    * When we sign in, we should make required api calls and pass into state
    * that way, we don't have to make multiple api calls on pages
    * on pages that need it, we will make more in-depth calls to get further data
-   * 
    * So, on successful sign in state will be like so:
    * state: {
    * signedIn,
@@ -34,8 +33,6 @@ const App = () => {
    *  ids
    * ]
    * }
-   * 
-   * 
    */
   useEffect(() => {
     client.on('connected', data => console.log('event happened', data))
@@ -65,13 +62,12 @@ const App = () => {
     client.emit('connected', {
       message: 'User connected'
     });
-    
    // eslint-disable-next-line react-hooks/exhaustive-deps
   },[]);
   const handleLogIn = (data) => {
     console.log(data);
     console.log('Logging In');
-    const { email, password } = data; 
+    const { email, password } = data;
     console.log(email);
     console.log(password);
     client.authenticate({
@@ -135,7 +131,7 @@ const App = () => {
                   The page you are looking for probably doesn't exist
                 </Route>
               </Switch>
-            </ul> 
+            </ul>
           </Suspense>
       </div>
       </Router>
